+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   directory; instead the malformed file is now removed
 - Fixed a small visual glitch on Linux where the selected file was not
   highlighted in the file list in the thin or expanded file manager modes
+- Fixed heading ID links not working with, e.g., accents
 
 ## Under the Hood